requirements.txt
pandas
numpy
matplotlib
flask==2.2.2
grpcio==1.47.0
grpcio-tools==1.47.0
scikit-learn